refactor(cli): extract unity assets handler
